% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{determine_class}
\alias{determine_class}
\title{Determine the class of text chunk}
\usage{
determine_class(text)
}
\arguments{
\item{text}{Text to process.}
}
\description{
Determine the class of text chunk
}
