% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{determine_class_one}
\alias{determine_class_one}
\title{Determine the class of text chunk}
\usage{
determine_class_one(text)
}
\arguments{
\item{text}{Text to process.}
}
\description{
Determine the class of text chunk
}
\keyword{internal}
