% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-paragraph.R
\name{find_code_boundaries}
\alias{find_code_boundaries}
\title{Find the start / stop of code lines}
\usage{
find_code_boundaries(body)
}
\arguments{
\item{body}{The body of a document, i.e. plain text.}
}
\description{
Find the start / stop of code lines
}
