% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-paragraph.R
\name{new_class_and_indent}
\alias{new_class_and_indent}
\title{List constructor
Helper function to construct a list with elements class and indent that
has sensible defaults.}
\usage{
new_class_and_indent(class = NA, indent = NA)
}
\arguments{
\item{class, indent}{Elements of the list to construct.}
}
\description{
List constructor
Helper function to construct a list with elements class and indent that
has sensible defaults.
}
\keyword{internal}
