% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{paragraph}
\alias{paragraph}
\alias{construct_paragraph}
\alias{deconstruct_paragraph}
\title{Tools for manipulating paragraphs}
\usage{
construct_paragraph(text, class = "ordinary text")

deconstruct_paragraph(text, element = "text")
}
\arguments{
\item{text}{The text of the paragraph.}

\item{class}{The class of the paragraph.}

\item{element}{Which element to select.}
}
\description{
A paragraph is a list with text and other elements.
}
\section{Functions}{
\itemize{
\item \code{construct_paragraph}: Constructs a paragraph.

\item \code{deconstruct_paragraph}: Exracts an element from a paragraph.
}}

\keyword{internal}
