% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{set_arg_write_tree}
\alias{set_arg_write_tree}
\title{Set the write_tree argument}
\usage{
set_arg_write_tree(write_tree)
}
\arguments{
\item{write_tree}{Whether or not to write tree.}
}
\description{
Sets the argument `write_tree` in [test_collection()] to be `TRUE` for R
versions higher or equal to 3.2, and `FALSE` otherwise since the second-level
dependency `DiagrammeR` from `data.tree` is not available for R < 3.2.
}
\keyword{internal}
