% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-paragraph.R
\name{split_text_into_header_and_body}
\alias{split_text_into_header_and_body}
\title{Split text into header and body}
\usage{
split_text_into_header_and_body(text, header = NULL)
}
\arguments{
\item{text}{Text to process.}

\item{header}{The header, in case there is one and it should be overwritten.}
}
\description{
Split text into header and body
}
