% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{split_text_into_paragraphs}
\alias{split_text_into_paragraphs}
\title{Split text into paragraphs}
\usage{
split_text_into_paragraphs(text, header = NULL)
}
\arguments{
\item{text}{Text to process.}

\item{header}{The header, in case there is one and it should be overwritten.}
}
\description{
Splits text into a list of elements, one corresponding to a paragraph,
for further processing. Each element
contains the text an the class of the chunks.
}
