% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stylermd.R
\docType{package}
\name{stylermd-package}
\alias{stylermd}
\alias{stylermd-package}
\title{Make a file or plain text fit 80 characters. Or any limit you want.}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\examples{
tidy_text(lorem_ipsum())
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lorenzwalthert/stylermd/issues}
  \item Report bugs at \url{https://github.com/lorenzwalthert/stylermd/issues}
}

}
\author{
\strong{Maintainer}: Lorenz Walthert \email{lorenz.walthert@icloud.com}

}
