% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{tidy_file}
\alias{tidy_file}
\title{Tidy a file}
\usage{
tidy_file(path, width = 80)
}
\arguments{
\item{path}{The path to the file.}

\item{width}{The maximal lenght of a line.}
}
\description{
Tidies a Markdown file
}
