% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_lines}
\alias{tidy_lines}
\title{Tidy lines}
\usage{
tidy_lines(text, add_only = FALSE)
}
\arguments{
\item{add_only}{If we should only add line breaks to lines that are too long
or if line breaks should also be removed.}
}
\description{
Tidyies mutliple line of code.
}
