% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_listing}
\alias{tidy_listing}
\title{Tidies a listing and indents correctly}
\usage{
tidy_listing(listing, width, spaces = c(0, 0))
}
\arguments{
\item{listing}{A listing, as described in the vignette "Data structures".}

\item{width}{The target width of the lines, passed to \code{\link[=tidy_lines]{tidy_lines()}}.}

\item{spaces}{Numeric vector of length two. Number of spaces for the
first line and all remaining lines.}
}
\description{
Indents the first line of a listing by \code{spaces_first} and the remaining
lines by \code{spaces_not_first} while maintaining the correct width.
}
\keyword{internal}
