% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut.R
\name{cut_long}
\alias{cut_long}
\title{Cut a string to a certain width}
\usage{
cut_long(text, width)
}
\arguments{
\item{text}{The text to cut.}

\item{width}{The maximal lenght of a line.}
}
\description{
Cut a string to a certain width
}
\keyword{internal}
