% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut.R
\name{cut_long}
\alias{cut_long}
\title{Cut a string to a certain length}
\usage{
cut_long(text, length = 80)
}
\arguments{
\item{text}{The text to cut.}

\item{length}{The maximal lenght of a line.}
}
\description{
Cut a string to a certain length
}
\keyword{internal}
