% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-paragraph.R
\name{determine_class}
\alias{determine_class}
\title{Determine the class of paragraphs}
\usage{
determine_class(paragraphs)
}
\arguments{
\item{paragraphs}{A list of non-header paragraphs.}
}
\description{
Determine the class of paragraphs
}
