% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine-class-of-paragraph.R
\name{determine_class_one}
\alias{determine_class_one}
\title{Determine the class of text chunk}
\usage{
determine_class_one(paragraph)
}
\arguments{
\item{paragraph}{A paragraph, i.e. a character vector, as defined in the
vignette "Data structures".}
}
\description{
Determine the class of text chunk
}
\keyword{internal}
