% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{tidy_file}
\alias{tidy_file}
\title{Tidy a file}
\usage{
tidy_file(path, width = getOption("stylermd.line_width"))
}
\arguments{
\item{path}{The path to the file.}

\item{width}{The maximal lenght of a line. If \code{length(width) > 1}, the
elements are used sequentially, one for each iteration. If there are more
iterations than values of \code{width}, the last value of \code{width} is used for
all remaining iterations.}
}
\description{
Tidies a Markdown file
}
