% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_line}
\alias{tidy_line}
\title{Tidy one line}
\usage{
tidy_line(text)
}
\arguments{
\item{text}{Text to prettify.}
}
\description{
Necessary to be called with \code{purrr::map(..., tidy_lines)} only if \code{add_only}
was set to \code{TRUE} because then we have multiple text lines to tidy.
}
\keyword{internal}
