% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_paragraph}
\alias{tidy_paragraph}
\title{Tidy up a paragraph}
\usage{
tidy_paragraph(paragraph)
}
\arguments{
\item{A}{paragraph to tidy.}
}
\description{
After text has been split into pargraphs with
[split_text_into_paragraphs()], this is the workhorse for actually formatting
the text.
}
