% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{tidy_text}
\alias{tidy_text}
\title{Tidy text}
\usage{
tidy_text(text, width = getOption("stylermd.line_width"))
}
\arguments{
\item{text}{The text to tidy.}

\item{width}{The maximal lenght of a line.}
}
\description{
Tidy text
}
